rafactor(rank-comparator):RankComparator를 Rank 클래스 내부로 이동
